Resize fitData.paramStartValues on init if needed.
When initializing fitData.paramNames is filled with data, according to
the model type. But paramStartValues isn't because it's expected to be
filled with data from the xml-file. But it only is, if there is data
in the xml-file. But when newly creating it, there's no such data.
Thus the sizes of the two lists would get out of sync and eventually
crash.