Always update the GSL parser from the y source
Add Bison as required build dependencies, and use it to always
regenerate at build time the GSL parser. This ensures that the parser
does not rot, that we benefit automatically from improvements in bison,
and there is no more need to rely on autogenerated sources added
statically among the others.

If the generate code gives warnings, they need to be fixed upstream in
bison, or worked around by disabling only for those sources the
compiler flags that cause the warnings.