Use KLocalizedString for attributeWarning
The result of i18n() with placeholders must have the placeholders
replaced as well, so calling QString("%1").arg(...) does not work, and
triggers warnings from KLocalizedString.

Convert all the attributeWarning instances to KLocalizedString,
replacing the missing attribute when needed.  As additional/related
change, remove the extra quoting from the reported attribute, as the
warning string already has it.