Switch to QPageSize for page size handling
Summary:
The current code has both a copy of page sizes from the Qt
implementation, and a list of page sizes (with own translations for
the items).

Switch the code to use QPageSize exclusively, with a single list of
sizes the only list needed; updatePaperSize() is greatly simplified
now.

Test Plan:
Builds fine, the list of page size is unchanged (but using
the UI names from QPageSize), and the page size handling is the same.

Reviewers: sgerlach, asemke

Reviewed By: asemke

Subscribers: kde-edu

Tags: #kde_edu

Differential Revision: https://phabricator.kde.org/D13731