[live data] Fix crash when trying to import non-ASCII/non-Binary data
Summary:
Only ASCII and Binary types are available, but the file may
be detected being another type. The file type combo box does
not include the automatically selected type and returns -1
for the currentIndex, which does not correspond to any file type.

The patch sets a default preview table and introduces a check
if a filter could be set.

Test Plan: - try to select a HDF5, ROOT, ... file in live data selection

Reviewers: #labplot, fkristof

Reviewed By: fkristof

Subscribers: kde-edu

Tags: #kde_edu

Differential Revision: https://phabricator.kde.org/D14500