[ROOT] Extend ROOT importer by possibility to load trees and tuples
Summary:
ROOT trees allow the storage of associated data of different types
in so-called branches. Each branch contains at least one leaf which
labels the contained data. The actual data is stored in baskets,
which contain data of predefined size.
This patch extends the ROOTHist class by the possibility to import
such trees and ntuples, which are a easier-to-use version of simple
trees.

Further changes:
- fix loading of large files
- load streamer info from file, use a default description if none is found

Test Plan: - import test files

Reviewers: #labplot, asemke

Reviewed By: #labplot, asemke

Subscribers: asemke, yurchor, kde-edu

Tags: #kde_edu

Differential Revision: https://phabricator.kde.org/D17610