give netcdf a finder
Summary:
netcdf may be built with either cmake or autotools. as it turns out the
autotools install will not have a cmake config, so finding it exclusively
via the config is not possible.

instead use a finder which wraps a find_package(config) call and fall
back to manual finding iff the config isn't found.

the finder defines the same target as the config would, making them
interchangeable.

Test Plan: builds with both operation modes

Reviewers: sgerlach

Reviewed By: sgerlach

Subscribers: kde-edu

Tags: #kde_edu

Differential Revision: https://phabricator.kde.org/D20891