Select curve only, when clicked near curve.
The current behavior is, that when you click into a cartesianPlot with a curve, it is not possible to call the contextmenu of the cartesianPlot or select a curve, which is behind the most in front curve, because the clicks and mouse hovers gets the most in front object and for a curve the complete boundingrect will be used. With this patch it will be looked, if the mousecursor is near to the curve points and only a action on the curve will occur, when the distance to the curve is lower than a factor.

Differential Revision: https://phabricator.kde.org/D17731