Fix updating live data "On New Data"
If a watched file is updated by being replaced, the containing
directory is watched now until the file is recreated.
Also a timeout is added to wait until the file is entirely written.

This fixes losing the connection to the file, which lead to not
updating the linked contents any longer.

While being paused, eventual file updates are registered and
read once the user decides to continue reading.
Also fix the behavior after a manual update during a paused
periodic reading.