When changing axis properties in a QLineEdit, don't process the changed signal, don't re-set the same value again in the text field that was just entered.
For example, when changing the scaling factor for the axis in AxisDock,
the slot AxisDock::scalingFactorChanged() is called which calls
axis->setScalingFactor(scalingFactor). In Axis::setScalingFactor() the
changed signal is sent with the new value. This signal is processed in
AxisDock::axisScalingFactorChanged() where we set the same value again
in QLineEdit. This sets the current position of the cursor to the end of
the text field blocking the user from entering digits one after another.