Don't crash when doing drag&drop between different main windows.
When starting the drag, we serialize the pointers of the
selected/dragged objects to the mime data. When deserializing this data
in the drop event in a different main window this results in invalid
pointers and leads to a crash. To avoid this, we also serialize and
deserialize the point of the pjoject object and add additional check for
it - if the project pointers in the dropped data differ from the current
project we don't accept the drop. The actual feature itself - drag&drop
between different main windows - is interesting and should be supported
in futur but will required a totally different serialization/deserialization
of the objects.

BUG: 428698
FIXED-IN: 2.8.1