Fixed crashes caused by the wrong results in AspectTreeModel::modelIndexOfAspect().
In c949e15 AbstractAspect::children(AspectType, ChildIndexFlags) was changed to check the exact aspect type instead of the inheritance hierarchy. 
This lead to wrong results give the current usage of both AbstractAspect::children() functions (templated one and via AspectType). The logic in the second function
was corrected and is now similar to the templated function checking the inheritance path with dynamic_cast.
Furthermore, some more cleanup in usage of AbstractAspect::children(), prefer the templated version of it.