[parser] use the right type for characters
Store each read character to be parsed as int, and not as char; the
difference is actually visible in architectures (e.g. PPC, s390x) that
have char as unsigned type, triggering the compiler warnings when
comparing char to EOF:

  backend/gsl/parser.y: In function 'yylex':
  backend/gsl/parser.y:334:8: warning: comparison is always false due to limited range of data type [-Wtype-limits]
  backend/gsl/parser.y:420:12: warning: comparison is always true due to limited range of data type [-Wtype-limits]
  backend/gsl/parser.y:423:9: warning: comparison is always true due to limited range of data type [-Wtype-limits]

EOF is defined according to the architecture, so comparing with a type
that has a different size and signedness leads to problems; in
particular, the parser is basically broken on those architectures, and
the "fittest" test shows this failure.

The right way is to store characters as int, as also described in the
documentation of getchar(), not used here that however references EOF.

This partially reverts commit 68f4d1911164472bf23bbdbf11db2640b0e6c12a
and commit e2bbf36764c1ffadb6d2de492f3b5f6cf57acd0b, which went to fix
code to the wrong direction.

(cherry picked from commit d4c2da6fbdf9e389c0195edc8b787dcbcd99c3e5)