Properly read the example project files that are in the old format compressed with Gzip.
Even though we we're delivering the example projects starting with 2.9 where Xz is used for the compression, the users still can decide to add their own project files in the old format (using Gzip) to the folder where the example projects are located. Without any handling for Gzip the current implementation resutls in an endless loop. To make the code more robust here, we add the support for the old format, too.