[worksheet] removed the wait cursor when updating/re-calculating curves.
With multiple curves on the plot, retransforming the plot ranges (zooming, etc.) triggers retranform in the curves
whith a sequence of WAIT_CURSOR and RESET_CURSOR. This behavior is annoying and flickering and is also slowing
down the actual updates in the curves. This commit changes this behavior. In case we still decide later to show a wait cursor
when navigating with expensive re-calculate's in the curves, we need to handle the status of the cursor in the plot and 
not in each curve on the plot to avoid flickering.