[spreadsheet] the fix done in d4c61765d657b6ad4dfc4db3c082a16e7cdb954d has caused test failures on windows and freebsd which are not clear and which don't happen on linux. Reworking the fix trying to adress the test failures.
Changed the logic to create SpreadsheetModel to the previous state where the model was created in the SpreadsheetView and only moved the deletion of it to the desctructor of Spreadsheet.
The model is only needed in the view and it makes sense to create it where and only when it's required. The ownership and the deletion of model is transferred to Spreadsheet - if the view is deleted
or if we want to support multiple views for one Spreadsheet in future, there is no need to create the model again, so we keep it in Spreadsheet.