[worksheet] don't crash when undoing the step "add info element".
When adding an info element, it's text label is modified and we have multiple undo entries within the undo macro.
Undoing this macro triggers also the undo of the text modification which on the other hand triggers InfoElement::retransform().
In the retransform we don't have any parent item anymore (the scene item was already deleted) and we need to handle this case.