[spreadsheet] reduce the memory consumption during the spreadsheet export to Sqlite by inserting the data into the table in multiple chunks.
When exporting the spreadsheet to Sqlite, internally a string for the INSERT-statement is created to insert the values into the database table.
When dealing with big amount of data in the spreadhseet (multiple millions of cells). this string can become very big which also causes a very high memory consumption.
In the worst case we're running out of memory and the process is killed. To solve this problem we insert the data now in multiple chunks having 10k spreadsheet rows maximum.