[sql import] reduce the memory consumption when importing from a table.
When reading from a table, the total number of rows to be read can be determined
via a SELECT COUNT(*) statement in a separate query and don't need to navigate back and forth in the main query.
So, in this case we can use QSqlQuery::setForwardOnly() to reduce the memory consumption.