[worksheet] 	don't crash when deleting a text label that was added after worksheet objects (plots and curves) were created via the plot data dialog.
When a text label is added after new curves were created via PlotDataDialog,  
the undo or delete steps for this label lead to a crash in the handling of graphics effects in Qt. 
The root cause for the crash is not completely clear (maybe related ot labels child ScaledTextItem)
so we deactivate the fade-in effect completely for text labels.

BUG: 455096
FIXED-IN: 2.10