[datasets] set the default value for the ASCI filter parameter "skip empty parts" to false.
All R datasets have an empty column at the beginning at skipping the empty parts wrongly imports this data.
We only need to skip the empty parts for the JSE data collection and it's easier to change the default value
and to add the parameter to its JSON file then to do the same for R's datasets JSON file.