[worksheet] use system's default font in text labels insteads of hard-coded Arial 12pt.
Since Arial 12 in plot's title doesn't look nice in combination with system's default font used for axis labels, our default plot doesn't look nice neither.
The switch to Arial 12pt was done in 7a7e2242113b46a74e41e14e358032dc8e52b70b because of BUG 451062 which was not reproducible for us on any of the supported
platforms and was because of some weird font issues on windows in Thailand. Since this is a very rare problem and the user still can fix it by selecting a differnt font,
we should go for the default font everywhere which is fine for the majority of our users.

fix #209