[ascii import] when reading data,  assume it's UTF8 encoded and not Latin1.
UTF8 is the commonly used encoding nowadays.  We used it in the past and switched to Latin1 because of performance optimizations
which was clearly a very bad decision. We go back to UTF8 now.  This won't solve the problems with UTF16 as reported in
https://invent.kde.org/education/labplot/-/issues/541 and probably other encodings that are not subset of UTF8 but this needs to be
addressed differently if we want to support them properly in future.

BUG: 470338
FIXED-IN: 2.10.1

Added the actual changes in AsciiFilter.cpp that were forgotten in the previous commit.