[import] change the default value of the header line in AsciiFilter from 0 to 1 to fix the import of datasets where the exact header line is not specified.
In the JSON files for the definitions of the datasets the field "use_first_row_for_vectorname"
is used to determine whether the header has to be used or not. With the change
in 2.10 additionally the header line was introduce that needs to be specified.
The default value in AsciiFilterPrivate.h is 0, but in the AsciiWidgetOption
this value is being overwritten with 1 so we properly read the column names
from the first line in the file on default. So, we change it to 1 now to fix the datasets.

filter.setHeaderEnabled(true) was activated again in the tests even though
the default value is already 'true' to make explizit what we're reading in every test.