Reduce the number of range updates in the plot area and retranfsforms in its children.
If the size of the plots that don't plot the data points directly like XYCurve
and need to recalculate their internal structures on data changes like BoxPlot,
Histogram, etc., has changed because of the actual data changes or because of
new plot settings, emit dataChanged() in order to recalculate the data ranges
in the parent plot area and to retransform all its children. Just call retransform()
to update the plot only if the ranges didn't change.