Disconnect when removing object or reseting the column
When an object will be removed, all linked objects set their member to a nullptr. If the signal gets not disconnected, for example when the deleted object changes it's name the linked object gets the signal and tries to update the path. In this case the linked object tries to access the member of the deleted object, but it is set to a nullptr and then labplot crashes