Remove all worksheet links when project gets closed
Description: When closing the project, the destructor will be called, all childs get deleted without sending the aspectAboutToBeRemoved() signal (see AbstractAspect desctructor). Therefore the PreviewWidget still contains the old worksheet pointer which are anymore valid. When creating a new project the WorksheetPreviewWidget will be shown and tries to update the worksheet previews. Because of this it crashes