Only show warnings collected during the project load if the debug tracing is active.
Warning are not really helpful in most of the cases and contain mostly the information
about missing and default values being used, especially when loading older projects.
In order not to pollute the output in the terminal on project load, we don't
shown any warnings on default. Errors are still shown.