check Range for every curve
Description: If one curves starts at zero for a log10 scaling, the initial value cannot be reached, so a fixed value of 0.01 is used. If the second curve goes down to 0.00001 the startvalue for the autoscale is still 0.01 because the minimum after the loop and before checkRange is 0. After applying the recommandations of checkRange, the start value is hardcoded to 0.01.