Refactored the handling of the dock widgets and their states.
We add the standard dock widgets to our initial default and hard-coded positions first.
if the user has changed the position of these docks or saved and restored them from the project file,
we put the default docks to our initial/default position when a new project is created or a project
is opened where the state of the default docks was not saved.
So, we always work with the same initial position of the docks and are not able to save the user-defined
settings - this will be implemented later.