Fix crash
Reason/Description: If a column in the spreadsheet gets remove, because of an import. It can happen, that MainWindow activates a different dockwidget, which leads to a redraw of the Spreadsheetview. During this import, the spreadsheetmodel m_columnCount and the spreadsheet column count are not in sync
Files: SpreadsheetModel.cpp