Improving parser speed by removing all symbols which are not required for the second parse
Reason/Description: In the first parse, the symbols are determined which are relevant for parsing the current line. For the next loop rounds we iterate only over the relevant symbols. So instead of looping over a few hundred symbols and comparing the name, we iterate only over a few (depending on the formula complexity, but less than 10 covers most cases)
Files: ColumnPrivate.cpp, ExpressionParser.cpp, ExpressionParser.h, and 16 more files