[axis] fix the initial ticks spacing after the switch from "number" to "spacing" ticks type.
Initially, the ticks spacing is not set (equal to zero) and after the switch to "spacing"
it's adjusted to the current spacing determined by the number of the ticks. Additionally to
the notification to the dock widget to show the newly calculated value we also
need to set the internal member variable so it's actually used when drawing the ticks.