Removed the action "close project" in the context menu of the project explorer.
There is no state "no project" anymore - we always have either an initial and default project and we load a project.
The action "close project" in the main menu was removed because of this and the similar action in the project explorer was forgotten.
This also fixes the problem with non-initialized project leading to crashes when the project is close and no new one was created/loaded yet
and we try to access it from other actions visible in the GUI like the data import.

BUG: 503800
FIXED-IN: 3.0