string extraction: exclude lib/examples/
lib/examples does not contain strings that should end up in the
message template (in fact it doesn't contain translatable strings
at the moment).
Also, it contains file with spaces which are not handled by the
current code of Messages.sh. While the extraction should be fixed
to account for spaces in file name, quickly workaround the issue
for now.