Fix crash in when an aspect gets removed which is not currently selected
Reason:
# Problem description: When having a seasonal composition with the STL Method and switching then to MSTL and back labplot crashes
## Steps to reproduce
1) Create a seasonal decomposition element and set all properties and columns
2) Select MSTL as method
3) Switch back to STL

## Expected
No crash

## Observed
crash, because SpreadsheetModel::handleAspectsRemoved() gets called before SpreadsheetModel::handleAspectsAboutToBeRemoved() is called

## Description
The problem is that during the creation of the seasonaldecomposition object, no connection between the result spreadsheet and the model exists, because the model does not exist yet (because the view was not needed yet which creates the model). During the switch from MSTL and STL a column gets removed. So the spreadsheet signals aspectsAboutToBeRemoved() but any view will react on it because any exists. Then the column gets removed which triggers a MainWin::handleAspectRemoved() call. Which automatically set the parent (The result spreadsheet) as current aspect and then the view (and therefore also the model) gets created and the connection is setup. After the column creation the Spreadsheet signals aspectsRemoved() and this time the previously created model handles the signal and we execute endRemoveColumns() before a beginRemoveColumns()